+++ tests/preset.rs
@@ -39,7 +39,7 @@ fn keys() {
 
         let actions = app.world().get::<Actions<Test>>(entity).unwrap();
         assert_eq!(
-            actions.action::<TestAction>().value(),
+            actions.value::<TestAction>().unwrap(),
             dir.into(),
             "`{key:?}` should result in `{dir}`"
         );
@@ -78,7 +78,7 @@ fn dpad() {
 
         let actions = app.world().get::<Actions<Test>>(ctx_entity).unwrap();
         assert_eq!(
-            actions.action::<TestAction>().value(),
+            actions.value::<TestAction>().unwrap(),
             dir.into(),
             "`{button:?}` should result in `{dir}`"
         );
@@ -117,7 +117,7 @@ fn sticks() {
 
             let actions = app.world().get::<Actions<Test>>(ctx_entity).unwrap();
             assert_eq!(
-                actions.action::<TestAction>().value(),
+                actions.value::<TestAction>().unwrap(),
                 dir.into(),
                 "`{axis:?}` should result in `{dir}`"
             );